Create unique filename for tests
The usage of mkstemp is awkward, as we are not interested in the created file,
only the filename. The reason for using mkstemp over tmpnam is the race
condition, as one is not guaranteed that the file is not created by a different
process between the creation from tmpnam till file creation. By closing and
removing the file right after mkstemp, we are pretty much re-introducing the
reasons for not using tmpnam..

But since clang is strict, we have to use a different function.